divide gridlike commands from element_commands
